+++ resources/views/components/lost-changes-warn.blade.php
@@ -19,6 +19,9 @@
             });
         }
 
+        // Tiptap
+        window.addEventListener('editor:dirty', () => isDirty = true);
+
         form.addEventListener('submit', () => {
             isDirty = false;
         });